Implement _.every function using reduce()
This one was very difficult to grasp as _.every is a higher order
function, and _.reduce is also a higher order function.

The key to this problem is making sure that once the iterator (the
callback function that aplies a truth test to every element in the
collection) returns false, the accumulator stays "false" until _.reduce
completes the iteration.

Finally, the return value of reduce (either true or false) is returned.