Implement _.delay function (see body)
In order to make this function work, use a Web API method called setTimeout(),
which takes a function as an argument to be executed after a delay in ms.

To access the follow-on arguments, we can process the array-like object
Function.arguments using properties such as the numerical index using [i] or
its length using .length.

The following chain of Array methods can also return a standard array:
Array.prototype.slice.call(arguments)

Then, slice the arguments array from the 3rd to the last element which will
serve as the arguments for the callback function in setTimeout().