Implement _.shuffle function (see body)
Algorithm is as follows:

Pick a random element from the array, using Math.random() to generate a
random index between 0 and the length of a copy of the original array.

Push the element into a new result array, adding elements from the back.
Repeat until there are no more elements to pick from our copy array.